+++ src/Controller/RootController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Controller;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\Routing\Annotation\Route;
+
+class RootController extends AbstractController
+{
+    #[Route('/', name: 'app_redirect')]
+    public function index(): RedirectResponse
+    {
+        return $this->redirectToRoute('api_doc');
+    }
+}