+++ CHANGELOG.md
@@ -18,6 +18,7 @@ See [keep a changelog] for information about writing changes to this log.
 - Added pagination and match filter
 - Added date filters
 - Ensure combined filters are possible
+- -Added api endpoinst for occurrences, location, tags, vocabularies and filters
 
 [keep a changelog]: https://keepachangelog.com/en/1.1.0/
 [unreleased]: https://github.com/itk-dev/event-database-imports/compare/main...develop